% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{attach_report}
\alias{attach_report}
\title{Set whether the output of the rendered R Markdown report should be attached to the email sent}
\usage{
attach_report(include_report = TRUE)
}
\arguments{
\item{include_report}{if TRUE, will include the output of the report.}
}
\description{
Set whether the output of the rendered R Markdown report should be attached to the email sent
}
\examples{
attach_report(TRUE) # Default is true will return the output of the report
attach_report(FALSE) # Will not include the output of the report
}
