% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_report_rendering_url}
\alias{get_report_rendering_url}
\title{Returns the URL for this specific rendered version of the report.}
\usage{
get_report_rendering_url()
}
\description{
Returns the URL for this specific rendered version of the report.
}
