% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{set_email_attachments}
\alias{set_email_attachments}
\title{Set the output files, and/or the email attachments created by this R Markdown report}
\usage{
set_email_attachments(output_files = NULL, email_attachments = NULL)
}
\arguments{
\item{output_files}{list of file paths of content that was created by the report}

\item{email_attachments}{list of file paths to the attachments created by the report}
}
\description{
Set the output files, and/or the email attachments created by this R Markdown report
}
