% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{set_email_metadata}
\alias{set_email_metadata}
\title{Configure the metadata associated with the email being sent.}
\usage{
set_email_metadata(email_subject = NULL, email_body_text = NULL,
  email_body_html = NULL, email_images = NULL)
}
\arguments{
\item{email_body_text}{String, if set, will override the plain text format of the email sent from Connect}

\item{email_body_html}{String, if set, will set the html body used in the email from Connect. @seealso blastula::compose_email}

\item{email_images}{base64 encoded image (or list of images).  It is easiest to compose these anoutput of the email$images from @seealso blastula::compose_email}

\item{emails_subject}{String, if set, will override the default subject from Connect}
}
\description{
Configure the metadata associated with the email being sent.
}
